support adaptive time steptting based on newton solver

+++ ogsFileIO/FemIO/ogs5/rf_pcs.h
@@ -89,7 +89,7 @@ public:
     double ExcavBeginCoordinate;          //WX
 
     size_t mesh_id;
-    size_t timegroup_id;
+    int timegroup_id;
 };
 
 bool PCSRead(const std::string&, std::vector<CRFProcess*> &pcs_vector);
